API_Database/requirements.txt
Flask==2.0.1
mysql-connector-python==8.0.26
PyJWT==2.1.0
requests==2.26.0
Flask-Cors==3.0.10
python-dotenv==0.19.0
